+++ CHANGELOG.md
@@ -6,5 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [1.0.0] - 2025-08-13
+
 ### Added
 
+- Adds oihana\exceptions\BindException
+- Adds oihana\exceptions\ExceptionTrait
+- Adds oihana\exceptions\MissingPassphraseException
+- Adds oihana\exceptions\ResponseException
+- Adds oihana\exceptions\UnsupportedOperationException
+- Adds oihana\exceptions\ValidationException
+-
+- Adds oihana\exceptions\http\Error403
+- Adds oihana\exceptions\http\Error404
+- Adds oihana\exceptions\http\Error500
+-